requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
cachetools==5.2.0
certifi==2022.9.24
greenlet==2.0.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
psycopg2==2.9.5
python-telegram-bot==20.0a4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
tornado==6.2
tzdata==2022.6
tzlocal==4.2
